Properties: Breakup module
`FileProperties` and `ChannelMask` are no longer in the root. They are moved to `lofty::properties`.
